# CRUDBooks/wwwroot/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "CRUD API Library",
    "version": "v1"
  },
  "paths": {
    "/registration": {
      "post": {
        "tags": ["Account"],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/login": {
      "post": {
        "tags": ["Account"],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/books": {
      "get": {
        "tags": ["Book"],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/book/{id}": {
      "get": {
        "tags": ["Book"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "put": {
        "tags": ["Book"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": ["Book"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/book/ISBN/{isbn}": {
      "get": {
        "tags": ["Book"],
        "parameters": [
          {
            "name": "isbn",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/book": {
      "post": {
        "tags": ["Book"],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "components": {}
}